order and orderitem functionality added

+++ src/main/java/io/murad/modern/ecommerce/service/OrderService.java
@@ -3,12 +3,16 @@ package io.murad.modern.ecommerce.service;
 import io.murad.modern.ecommerce.database.model.Order;
 import io.murad.modern.ecommerce.database.model.User;
 import io.murad.modern.ecommerce.dto.PlaceOrderDto;
+import io.murad.modern.ecommerce.exception.OrderNotFoundException;
 import io.murad.modern.ecommerce.repository.CartRepository;
 import io.murad.modern.ecommerce.repository.OrderItemRepository;
 import io.murad.modern.ecommerce.repository.OrderRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -34,7 +38,7 @@ public class OrderService {
         return orderList;
     }
 
-    public Order getOrder(int orderId) throws OrderNotFoundException {
+    public Order getOrder(Long orderId) throws OrderNotFoundException {
         Optional<Order> order = orderRepository.findById(orderId);
         if (order.isPresent()) {
             return order.get();
